refactor(server): boundary enforcement for the agent domain (C2)
Physical split delivered after all: a name-based guard cannot tell a
sanctioned agent-module import from a bypass (same names), so the module
path must be the choke point. lib/supabase/agentData.ts now holds the
agent-domain data access (settings, limits, rules, patterns, velocity,
daily stats, behavioral events, outcome learning — ~760 lines moved
verbatim), deliberately NOT re-exported from lib/supabase/index.ts.
db.ts keeps backend data access; getTelegramChatId now does its own
minimal user_settings query instead of the agent-domain accessor — the
E2 mixed-table wrinkle made explicit at the boundary.

lint:layering becomes scripts/lint-layering.sh with five rules: the A1
routes rule, the A2 raw-signing rule, and three new C2 rules —
lib/agent/signer importable only by src/agent/index.ts,
lib/supabase/agentData only by src/agent/, risk.js only by src/agent/
(+ agentData type import). All three negative-tested; clean tree passes.

server/src/agent/README.md documents the frozen interface, tiers, and
enforcement — the D-milestone job-payload surface, committed.

Implements #90
